+++ R/get_boards.R
@@ -1,7 +1,7 @@
 #' Get the discussion boards from a meetup group
 #'
 #' @template urlname
-#' @template api_key
+#' @template verbose
 #' @return A tibble with the following columns:
 #'    * id
 #'    * name
@@ -19,23 +19,23 @@
 #'@examples
 #' \dontrun{
 #' urlname <- "rladies-nashville"
-#' meetup_boards <- get_boards(urlname = urlname,
-#'                       api_key = api_key)
+#' meetup_boards <- get_boards(urlname = urlname)
 #'}
 #' @export
-get_boards <- function(urlname, api_key = NULL) {
-  api_method <- paste0(urlname, "/boards")
-  res <- .fetch_results(api_method, api_key)
+get_boards <- function(urlname,
+                       verbose = getOption("meetupr.verbose", rlang::is_interactive())) {
+  api_path <- paste0(urlname, "/boards")
+  res <- .fetch_results(api_path = api_path, verbose = verbose)
   tibble::tibble(
     id = purrr::map_int(res, "id"),
     name = purrr::map_chr(res, "name"),
     description = purrr::map_chr(res, "description"),
     created = .date_helper(purrr::map_dbl(res, "created")),
     updated = .date_helper(purrr::map_dbl(res, "updated")),
-    post_count = purrr::map_int(res, "post_count", .null = NA),
-    discussion_count = purrr::map_int(res, "discussion_count", .null = NA),
-    latest_reply_created = .date_helper(purrr::map_dbl(res, c("latest_reply", "created"), .null = NA)),
-    latest_reply_member_name = purrr::map_chr(res, c("latest_reply", "member", "name"), .null = NA),
+    post_count = purrr::map_int(res, "post_count", .default = NA),
+    discussion_count = purrr::map_int(res, "discussion_count", .default = NA),
+    latest_reply_created = .date_helper(purrr::map_dbl(res, c("latest_reply", "created"), .default = NA)),
+    latest_reply_member_name = purrr::map_chr(res, c("latest_reply", "member", "name"), .default = NA),
     resource = res
   )
 }