+++ man-roxygen/verbose.R
@@ -1,0 +1,2 @@
+#' @param verbose logical; do you want informative messages? `TRUE` by default in
+#' interactive sessions. Can be toggled by the `meetupr.verbose` option.