+++ tests/testthat/test-get_events.R
@@ -1,22 +1,49 @@
-context("get_events")
+expected_names <- c("id", "name", "created", "status", "time", "local_date", "duration", "local_time",
+                    "waitlist_count", "yes_rsvp_count", "venue_id", "venue_name",
+                    "venue_lat", "venue_lon", "venue_address_1", "venue_city", "venue_state",
+                    "venue_zip", "venue_country", "description", "link", "resource"
+)
 
-test_that("get_events() success case", {
-  meetup_events <- with_mock(
-    `httr::GET` = function(url, query, ...) {
-      load(here::here("tests/testdata/httr_get_get_events.rda"))
-      return(req)
-    },
-    meetup_events <- get_events(api_key="yay",
-                                urlname = "<3",
-                                event_status = "upcoming")
-  )
+test_that("get_events() works with one status", {
+  urlname <- "rladies-nashville"
+  vcr::use_cassette("get_events", {
+    past_events <- get_events(urlname = urlname,
+                              event_status = "past")
+  })
 
-  expect_equal(nrow(meetup_events), 1, label="check get_events() returns one result")
-  expect_equal(meetup_events$status, "upcoming", label="check get_events() content (status)")
+  expect_s3_class(past_events, "data.frame")
+  expect_true(
+    all(
+      names(past_events) == expected_names
+    ))
 })
 
-# TODO: multiple statuses
+test_that("get_events() works with multiple statuses", {
+  skip("Not working for now (the test, not the function)")
+  urlname <- "rladies-johannesburg"
+  vcr::use_cassette("get_events-2-status", {
+    past_events <- get_events(urlname = urlname,
+                              event_status = c("past", "upcoming"))
+  })
 
+  expect_s3_class(past_events, "data.frame")
+  expect_true(
+    all(
+      names(past_events) == expected_names
+    ))
+
+})
+
+test_that("get_events() has informative error messages", {
+  urlname <- "rladies-johannesburg"
+  expect_error(
+    get_events(urlname = urlname, event_status = "pasttt"),
+    "should be one of"
+    )
+  expect_error(
+    get_events(event_status = "past")
+  )
+})
 # TODO: event type is not allowed
 
 # TODO: "urlname is missing"