+++ tests/testthat/test-internals.R
@@ -1,17 +1,12 @@
-context("internals")
-
-test_that(".quick_fetch() success case", {
-  res <- with_mock(
-    `httr::GET` = function(url, query, ...) {
-      load(here::here("tests/testdata/httr_get_find_groups.rda"))
-      return(req)
-    },
-    res <- .quick_fetch(api_url = "fake url", # intentionally invalid as there is currently no validation
-                      api_key = "I <3 R-Ladies")
-    )
-
-  expect_equal(names(res), c("result", "headers"), info="check .quick_fetch() return value")
-  expect_equal(res$headers$`content-type`, "application/json;charset=utf-8", info="check .quick_fetch() header content-type")
+test_that("meetup_call() success case", {
+  vcr::use_cassette("quick_fetch", {
+    res <- meetup_call(
+      api_path = "rladies-nashville/events",
+      event_status = "past"
+      )
+  })
+  expect_equal(names(res), c("result", "headers"))
+  expect_equal(res$headers$`content-type`, "application/json; charset=utf-8")
 })
 
 # TODO .fetch_results()